[OPR-703]: Rearranges Procedure.Chain
Adds support for filter and flatMap
