+++ Sources/ProcedureKit/Transform.swift
@@ -51,53 +51,3 @@ open class AsyncTransformProcedure<Input, Output>: Procedure, InputProcedure, Ou
         transform(inputValue) { self.finish(withResult: $0) }
     }
 }
-
-public extension Procedure {
-
-    public struct Chain<Output> {
-        public let procedures: [Procedure]
-        public let tail: AnyOutputProcedure<Output>
-
-        internal func add<NewOutput>(newTail: AnyOutputProcedure<NewOutput>) -> Chain<NewOutput> {
-            var newProcedures = procedures
-            newProcedures.append(newTail)
-            return Procedure.Chain<NewOutput>(procedures: newProcedures, tail: newTail)
-        }
-    }
-}
-
-public extension Procedure.Chain {
-
-    func transform<TransformedOutput>(block: @escaping (Output) throws -> TransformedOutput) -> Procedure.Chain<TransformedOutput> {
-        let transform = TransformProcedure(transform: block)
-        transform.injectResult(from: tail)
-        return add(newTail: AnyOutputProcedure(transform))
-    }
-}
-
-public extension Procedure.Chain where Output: Sequence {
-
-    func reduce<U>(_ initial: U, nextPartialResult: @escaping (U, Output.Iterator.Element) throws -> U) -> Procedure.Chain<U> {
-        return add(newTail: AnyOutputProcedure(tail.reduce(initial, nextPartialResult: nextPartialResult)))
-    }
-
-    func map<U>(transform: @escaping (Output.Iterator.Element) throws -> U) -> Procedure.Chain<[U]> {
-        return add(newTail: AnyOutputProcedure(tail.map(transform: transform)))
-    }
-
-}
-
-public extension ProcedureQueue {
-
-    func add<Output>(chain: Procedure.Chain<Output>) {
-        add(operations: chain.procedures)
-    }
-}
-
-public extension OutputProcedure where Self: Procedure {
-
-    func transform<TransformedOutput>(block: @escaping (Output) throws -> TransformedOutput) -> Procedure.Chain<TransformedOutput> {
-        let tail = AnyOutputProcedure(self)
-        return Procedure.Chain(procedures: [tail], tail: tail).transform(block: block)
-    }
-}
