+++ Tests/ProcedureKitTests/TransformProcedureTests.swift
@@ -61,27 +61,3 @@ class AsyncTransformProcedureTests: ProcedureKitTestCase {
         XCTAssertProcedureFinishedWithErrors(timesTwo, count: 1)
     }
 }
-
-class ProcedureChainTests: ProcedureKitTestCase {
-
-    func test__all_procedures_in_chain_executes() {
-
-        // This method allows arbitrary chaining of blocks to
-        // any output procedure. This can be used to chain 
-        // together smaller units of work in a functional
-        // style 
-        let chain = NumbersProcedure()
-            .transform { $0.map { $0 * 2 } }
-            .map { $0 + 3 }
-            .reduce(0, nextPartialResult: +)
-
-        wait(for: chain)
-
-        // Numbers
-        XCTAssertProcedureFinishedWithoutErrors(chain.procedures[0])
-        XCTAssertProcedureFinishedWithoutErrors(chain.procedures[1])
-        XCTAssertProcedureFinishedWithoutErrors(chain.procedures[2])
-        guard let testResult = chain.tail.output.value?.value else { XCTFail("Did not get a result"); return }
-        XCTAssertEqual(testResult, 3+5+7+9+11+13+15+17+19+21)
-    }
-}
